Updated UBI minimal version and k8s versions supported by the driver

+++ pkg/controller/controller.go
@@ -23,6 +23,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"sort"
 	"strconv"
 
 	"github.com/container-storage-interface/spec/lib/go/csi"
@@ -338,6 +339,16 @@ func (s *Service) DeleteVolume(ctx context.Context, req *csi.DeleteVolumeRequest
 				listSnaps)
 		}
 
+		// Validate if filesystem has any NFS or SMB shares or snapshots attached
+		nfsExportResp, _ := arr.GetClient().GetNFSExportByFileSystemID(ctx, id)
+		if len(nfsExportResp.ROHosts) > 0 ||
+			len(nfsExportResp.RORootHosts) > 0 ||
+			len(nfsExportResp.RWHosts) > 0 ||
+			len(nfsExportResp.RWRootHosts) > 0 {
+			return nil, status.Errorf(codes.FailedPrecondition,
+				"filesystem %s cannot be deleted as it has associated NFS or SMB shares.",
+				id)
+		}
 		_, err = arr.GetClient().DeleteFS(ctx, id)
 		if err == nil {
 			return &csi.DeleteVolumeResponse{}, nil
@@ -348,6 +359,7 @@ func (s *Service) DeleteVolume(ctx context.Context, req *csi.DeleteVolumeRequest
 			}
 		}
 		return nil, err
+
 	} else if protocol == "scsi" {
 		// query volume groups?
 		vgs, err := arr.GetClient().GetVolumeGroupsByVolumeID(ctx, id)
@@ -535,11 +547,42 @@ func (s *Service) ControllerUnpublishVolume(ctx context.Context, req *csi.Contro
 				"failure checking nfs export status for volume unpublishing: %s", err.Error())
 		}
 
+		// we need to construct the payload dynamically otherwise 400 error will be thrown
+		var modifyHostPayload gopowerstore.NFSExportModify
+		sort.Strings(export.ROHosts)
+		index := sort.SearchStrings(export.ROHosts, ip)
+		if len(export.ROHosts) > 0 {
+			if index >= 0 {
+				modifyHostPayload.RemoveROHosts = []string{ip + "/255.255.255.255"} // we can't remove without netmask
+			}
+		}
+
+		sort.Strings(export.RORootHosts)
+		index = sort.SearchStrings(export.RORootHosts, ip)
+		if len(export.RORootHosts) > 0 {
+			if index >= 0 {
+				modifyHostPayload.RemoveRORootHosts = []string{ip + "/255.255.255.255"} // we can't remove without netmask
+			}
+		}
+
+		sort.Strings(export.RWHosts)
+		index = sort.SearchStrings(export.RWHosts, ip)
+		if len(export.RWHosts) > 0 {
+			if index >= 0 {
+				modifyHostPayload.RemoveRWHosts = []string{ip + "/255.255.255.255"} // we can't remove without netmask
+			}
+		}
+
+		sort.Strings(export.RWRootHosts)
+		index = sort.SearchStrings(export.RWRootHosts, ip)
+		if len(export.RWRootHosts) > 0 {
+			if index >= 0 {
+				modifyHostPayload.RemoveRWRootHosts = []string{ip + "/255.255.255.255"} // we can't remove without netmask
+			}
+		}
+
 		// Detach host from nfs export
-		_, err = arr.GetClient().ModifyNFSExport(ctx, &gopowerstore.NFSExportModify{
-			RemoveRWRootHosts: []string{ip + "/255.255.255.255"}, // You can't remove without netmask
-			RemoveRWHosts:     []string{ip + "/255.255.255.255"},
-		}, export.ID)
+		_, err = arr.GetClient().ModifyNFSExport(ctx, &modifyHostPayload, export.ID)
 		if err != nil {
 			if apiError, ok := err.(gopowerstore.APIError); !(ok && apiError.HostAlreadyRemovedFromNFSExport()) {
 				return nil, status.Errorf(codes.Internal,
@@ -577,7 +620,12 @@ func (s *Service) ValidateVolumeCapabilities(ctx context.Context, req *csi.Valid
 			supported = false
 			reason = ErrUnknownAccessMode
 			break
+		// SINGLE_NODE_WRITER to be deprecated in future
 		case csi.VolumeCapability_AccessMode_SINGLE_NODE_WRITER:
+			break
+		case csi.VolumeCapability_AccessMode_SINGLE_NODE_SINGLE_WRITER:
+			break
+		case csi.VolumeCapability_AccessMode_SINGLE_NODE_MULTI_WRITER:
 			break
 		case csi.VolumeCapability_AccessMode_SINGLE_NODE_READER_ONLY:
 			break
@@ -973,46 +1021,52 @@ func (s *Service) ControllerGetVolume(ctx context.Context, req *csi.ControllerGe
 		// check if filesystem exists
 		fs, err := s.Arrays()[arrayID].Client.GetFS(ctx, id)
 		if err != nil {
-			return nil, status.Errorf(codes.NotFound, "failed to find filesystem: %s with error: %v", id, err.Error())
-		}
-
-		// get exports for filesystem if exists
-		nfsExport, err := s.Arrays()[arrayID].Client.GetNFSExportByFileSystemID(ctx, fs.ID)
-		if err != nil {
 			if apiError, ok := err.(gopowerstore.APIError); !ok || !apiError.NotFound() {
-				return nil, status.Errorf(codes.NotFound, "failed to find nfs export for filesystem with error: %v", err.Error())
+				return nil, status.Errorf(codes.NotFound, "failed to find filesystem %s with error: %v", id, err.Error())
 			}
+			abnormal = true
+			message = fmt.Sprintf("Filesystem %s is not found", id)
 		} else {
-			// get hosts publish to export
-			hosts = append(nfsExport.ROHosts, nfsExport.RORootHosts...)
-			hosts = append(hosts, nfsExport.RWHosts...)
-			hosts = append(hosts, nfsExport.RWRootHosts...)
+			// get exports for filesystem if exists
+			nfsExport, err := s.Arrays()[arrayID].Client.GetNFSExportByFileSystemID(ctx, fs.ID)
+			if err != nil {
+				if apiError, ok := err.(gopowerstore.APIError); !ok || !apiError.NotFound() {
+					return nil, status.Errorf(codes.NotFound, "failed to find nfs export for filesystem with error: %v", err.Error())
+				}
+			} else {
+				// get hosts publish to export
+				hosts = append(nfsExport.ROHosts, nfsExport.RORootHosts...)
+				hosts = append(hosts, nfsExport.RWHosts...)
+				hosts = append(hosts, nfsExport.RWRootHosts...)
+			}
 		}
-
 	} else {
 		// check if volume exists
 		vol, err := s.Arrays()[arrayID].Client.GetVolume(ctx, id)
 		if err != nil {
-			return nil, status.Errorf(codes.NotFound, "failed to find volume: %s with error: %v", id, err.Error())
-		}
-
-		// get hosts published to volume
-		hostMappings, err := s.Arrays()[arrayID].Client.GetHostVolumeMappingByVolumeID(ctx, id)
-		if err != nil {
-			return nil, status.Errorf(codes.NotFound, "failed to get host volume mapping for volume: %s with error: %v", id, err.Error())
-		}
-		for _, hostMapping := range hostMappings {
-			host, err := s.Arrays()[arrayID].Client.GetHost(ctx, hostMapping.HostID)
-			if err != nil {
-				return nil, status.Errorf(codes.NotFound, "failed to get host: %s with error: %v", hostMapping.HostID, err.Error())
+			if apiError, ok := err.(gopowerstore.APIError); !ok || !apiError.NotFound() {
+				return nil, status.Errorf(codes.NotFound, "failed to find volume %s with error: %v", id, err.Error())
 			}
-			hosts = append(hosts, host.Name)
-		}
-
-		// check if volume is in ready state
-		if vol.State != gopowerstore.VolumeStateEnumReady {
 			abnormal = true
-			message = fmt.Sprintf("Volume is in %s state", string(vol.State))
+			message = fmt.Sprintf("Volume %s is not found", id)
+		} else {
+			// get hosts published to volume
+			hostMappings, err := s.Arrays()[arrayID].Client.GetHostVolumeMappingByVolumeID(ctx, id)
+			if err != nil {
+				return nil, status.Errorf(codes.NotFound, "failed to get host volume mapping for volume: %s with error: %v", id, err.Error())
+			}
+			for _, hostMapping := range hostMappings {
+				host, err := s.Arrays()[arrayID].Client.GetHost(ctx, hostMapping.HostID)
+				if err != nil {
+					return nil, status.Errorf(codes.NotFound, "failed to get host: %s with error: %v", hostMapping.HostID, err.Error())
+				}
+				hosts = append(hosts, host.Name)
+			}
+			// check if volume is in ready state
+			if vol.State != gopowerstore.VolumeStateEnumReady {
+				abnormal = true
+				message = fmt.Sprintf("Volume is in %s state", string(vol.State))
+			}
 		}
 	}
